Add password update functionality

+++ routes/userPages.js
@@ -1,6 +1,8 @@
 // route for user specific pages or actions
 const express = require('express');
 const router = express.Router();
+const bcrypt = require('bcrypt');
+
 const authenticateToken = require('../tokenAuthMiddleware');
 const User = require('../models/user');
 
@@ -38,15 +40,41 @@ router.put('/updateUsername', authenticateToken, async (req, res) => {
 router.delete('/deleteAccount', authenticateToken, async (req, res) => {
   try {
     const userData = await User.findOne({ _id: req.userId });
+    const { password } = req.body;
 
     if (userData) {
-      await User.deleteOne({ _id: req.userId });
-      return res.status(200).json({ message: 'Account deleted successfully' });
+      const isPasswordMatch = await bcrypt.compare(password, userData.password);
+      if (isPasswordMatch) {
+        await User.deleteOne({ _id: req.userId });
+        return res.status(200).json({ message: 'Account deleted successfully' });
+      }
+      return res.status(401).json({ message: 'Invalid password' });
     }
     return res.status(404).json({ message: 'No data found' });
   } catch (error) {
     console.error('Error querying database', error);
     return res.status(500).json({ message: 'Error deleting account' });
+  }
+});
+
+router.put('/updatePassword', authenticateToken, async (req, res) => {
+  try {
+    const { currentPassword, newPassword } = req.body;
+    const userData = await User.findOne({ _id: req.userId });
+
+    if (userData) {
+      const isPasswordMatch = await bcrypt.compare(currentPassword, userData.password);
+      if (isPasswordMatch) {
+        userData.password = await bcrypt.hash(newPassword, 10);
+        await userData.save();
+        return res.status(200).json({ message: 'Password updated successfully' });
+      }
+      return res.status(500).json({ message: 'Invalid password' });
+    }
+    return res.status(404).json({ message: 'No data found' });
+  } catch (error) {
+    console.error('Error querying database', error);
+    return res.status(500).json({ message: 'Error updating password' });
   }
 });
